Fix problem with releasing domains
The way domains are being used for contexts, etc.
we need a refcnt as pointers to dom's are turned
to the app, which then is suppose to free them.
At the same time, dom's are also used internally
in SOS.  This kind of thing needs to use ref counts
to avoid possible multiple attempts to free underlying
resources associated with a dom struct.

With this commit, the ctx test now passes using the
GNI provider.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>